requirements.txt
python-telegram-bot==20.7
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2
gunicorn==21.2.0
flask==3.0.0
jinja2==3.1.2